Flatten comparison strategy by merging the comparison strategy interface, the abstract comparison class and the parallel implementation. Also replace token mapper factory method via constructor call.
